+++ ccm-shortcuts/src/test/resources-wildfly-remote-h2-mem/scripts/create_ccm_shortcuts_schema.sql
@@ -585,10 +585,10 @@ DROP SEQUENCE IF EXISTS hibernate_sequence;
     create table CCM_CORE.STYLE_MEDIA_QUERIES (
         MEDIA_QUERY_ID bigint not null,
         MAX_WIDTH_UNIT varchar(255),
-        MAX_WIDTH_VALUE float,
+        MAX_WIDTH_VALUE double,
         MEDIA_TYPE varchar(255),
         MIN_WIDTH_UNIT varchar(255),
-        MIN_WIDTH_VALUE float,
+        MIN_WIDTH_VALUE double,
         primary key (MEDIA_QUERY_ID)
     );
